v2.8

+++ backend/controllers/PostController.php
@@ -2,10 +2,12 @@
 
 namespace backend\controllers;
 
+use Yii;
 use common\models\Post;
 use common\models\PostSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
 
 /**
@@ -67,6 +69,9 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+        if (!Yii::$app->user->can('createPost')) {
+            throw new ForbiddenHttpException("对不起，你没有进行该操作的权限。");
+        }
         $model = new Post();
 
         if ($this->request->isPost) {
@@ -91,6 +96,9 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->can('updatePost')) {
+            throw new ForbiddenHttpException("对不起，你没有进行该操作的权限。");
+        }
         $model = $this->findModel($id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -111,6 +119,9 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->can('deletePost')) {
+            throw new ForbiddenHttpException("对不起，你没有进行该操作的权限。");
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
